repurpose dependOslModule
Previously it accepted a module name and base version, found the full
version for the project's minecraft version, and added a dependency for
that version.
Now, it accepts an OSL version and a module name, finds the module base
version for the OSL version, then finds the full version for the
project's minecraft version, and adds a dependency for that version.